Fixed typo, added css anims
Fixed and improved some wordings to add more meaning while entering.
Added placeholders for the boxes for user to understand better what to input without any confusion.
Also added appropriate margin and padding between the fields to make the webpage look clean and nice.
Added a linear color gradient to further enhance the looks and aesthetics of the webpage.
Added a snowfall effect using css animation and wrote the js code for the working.